afaik, this makes all the units work.
the results for rigid block are correct.
for [de]coupled, the output is the same as the program given to us. however, if i convert all the parameters into the new units and run it again, the results are not the expected (ft -> cm) results. so...half of it works?

because of the difference in precision of ft and cm, make a local DecimalFormat to keep correct formatting.
fix and add some unit labels.
other constants and fixes, too