updates from the hard drive erasure, and other stuff
fix rigid block dual slope in english. the problem was actually a double comparison to zero. one was being done to .0001, the other to exactly zero. change both to see if they are less than 1e-15. this gives up better answers and fixes the metric/english discrepancy.

also commit the start of untested code to add PGV, which is coming with more cleaner init DB code.